Measure a building's stage in two bytes, not four
A placed building never appeared. Its branch of the unit state names the
handle, the design and the construction stage, and the stage was going out
four bytes wide where the client reads two - so everything behind it in the
message arrived two bytes late and the client had no building to show. The
branch is nine bytes, not eleven.

The same two bytes were wrong at the front of the build-progress message,
where the handle led as four. There the damage was worse: the model the
client compares against what it is currently drawing was read out of the
middle of two other numbers, so a stage that changed could not change the
model.

Also puts the affiliations back in the order the client is known to read
them, faction first. Moving the faction to the third slot left every unit
reading as no faction at all, which is one side for everyone, and drops an
opcode that was declared, never sent, and sat on a number already spoken
for.

+++ AAEmu.Game/Core/Packets/G2C/SCHouseBuildProgressPacket.cs
@@ -3,14 +3,26 @@ using AAEmu.Game.Core.Network.Game;
 
 namespace AAEmu.Game.Core.Packets.G2C;
 
+/// <summary>
+/// How far along a building is, and what it should look like while it gets there.
+/// </summary>
+/// <remarks>
+/// Fourteen bytes. The handle leads at two - it was going out as four, which pushed the three
+/// fields behind it along by two bytes each and left the client reading the model out of the
+/// middle of two numbers.
+///
+/// The client finds the building by the handle, compares the model with the one it is currently
+/// drawing, and swaps the model when they differ; the two step counts it just stores. So the model
+/// is the one belonging to the stage the building is at, not the finished design's.
+/// </remarks>
 public class SCHouseBuildProgressPacket : GamePacket
 {
-    private readonly uint _tl;
+    private readonly ushort _tl;
     private readonly uint _modelId;
     private readonly int _allStep;
     private readonly int _curStep;
 
-    public SCHouseBuildProgressPacket(uint tl, uint modelId, int allStep, int curStep) : base(SCOffsets.SCHouseBuildProgressPacket, 5)
+    public SCHouseBuildProgressPacket(ushort tl, uint modelId, int allStep, int curStep) : base(SCOffsets.SCHouseBuildProgressPacket, 5)
     {
         _tl = tl;
         _modelId = modelId;
@@ -20,10 +32,10 @@ public class SCHouseBuildProgressPacket : GamePacket
 
     public override PacketStream Write(PacketStream stream)
     {
-        stream.Write(_tl);
-        stream.Write(_modelId);
-        stream.Write(_allStep);
-        stream.Write(_curStep);
+        stream.Write(_tl);      // tl      : u16
+        stream.Write(_modelId); // modelId : u32
+        stream.Write(_allStep); // allstep : i32
+        stream.Write(_curStep); // curstep : i32
         return stream;
     }
 }
